Extract provider logic from resource client
