+++ plugins/terraform_resource.py
@@ -21,7 +21,14 @@ import json
 import os
 import tempfile
 from pathlib import Path
+from typing import Optional
 
+from inmanta.agent import config
+from inmanta.agent.agent import AgentInstance
+from inmanta.agent.handler import CRUDHandler, HandlerContext, ResourcePurged, provider
+from inmanta.agent.io.local import IOBase
+from inmanta.protocol.endpoints import Client
+from inmanta.resources import Id, PurgeableResource, resource
 from inmanta_plugins.terraform.helpers.const import TERRAFORM_RESOURCE_STATE_PARAMETER
 from inmanta_plugins.terraform.helpers.param_client import ParamClient
 from inmanta_plugins.terraform.helpers.utils import (
@@ -35,13 +42,7 @@ from inmanta_plugins.terraform.tf.terraform_provider_installer import ProviderIn
 from inmanta_plugins.terraform.tf.terraform_resource_client import (
     TerraformResourceClient,
 )
-
-from inmanta.agent import config
-from inmanta.agent.agent import AgentInstance
-from inmanta.agent.handler import CRUDHandler, HandlerContext, ResourcePurged, provider
-from inmanta.agent.io.local import IOBase
-from inmanta.protocol.endpoints import Client
-from inmanta.resources import Id, PurgeableResource, resource
+from plugins.tf.terraform_provider import TerraformProvider
 
 
 @resource(
@@ -108,9 +109,17 @@ class TerraformResource(PurgeableResource):
 class TerraformResourceHandler(CRUDHandler):
     def __init__(self, agent: "AgentInstance", io: "IOBase") -> None:
         super().__init__(agent, io=io)
-        self.resource_client = None
+        self.provider: Optional[TerraformProvider] = None
+        self._resource_client: Optional[TerraformResourceClient] = None
         self.log_file_path = ""
         self.private_file_path = ""
+
+    @property
+    def resource_client(self) -> TerraformResourceClient:
+        if self._resource_client is None:
+            raise RuntimeError("The resource client is not setup")
+
+        return self._resource_client
 
     def _agent_state_dir(self, resource: TerraformResource) -> str:
         # Files used by the handler should go in state_dir/cache/<module-name>/<agent-id>/
@@ -243,12 +252,18 @@ class TerraformResourceHandler(CRUDHandler):
             param_client=param_client,
         )
 
-        self.resource_client = TerraformResourceClient(
+        self.provider = TerraformProvider(
             binary_path,
             self.log_file_path,
-            terraform_resource_state,
         )
-        self.resource_client.open(resource.provider_config)
+        self.provider.open()
+        self.provider.configure(resource.provider_config)
+
+        self._resource_client = TerraformResourceClient(
+            self.provider,
+            terraform_resource_state,
+            ctx.logger,
+        )
 
         resource.resource_config = build_resource_state(
             resource.resource_config,
@@ -262,7 +277,9 @@ class TerraformResourceHandler(CRUDHandler):
          - Stop the provider process
          - Cleanup the logs
         """
-        self.resource_client.close()
+        if self.provider is not None:
+            self.provider.close()
+
         with open(self.log_file_path, "r") as f:
             lines = f.readlines()
             if lines:
